Moved common client/server functions to common library and added timestamp function.
