+++ GroundControl/main.cpp
@@ -7,70 +7,15 @@
 #include <chrono>
 #include <sstream>
 #include <iomanip>
-#include <WinSock2.h>
-#include "DataPacket.h"
-
 #include "Common.h"
+using namespace common;
 
-#pragma comment(lib, "ws2_32.lib")
 const int PORT = 8080;
-
-uint32_t calculateChecksum(const DataPacket& packet) {
-
-    const uint8_t* bytes = reinterpret_cast<const uint8_t*>(&packet);
-    uint32_t checksum = 0;
-    for (size_t i = 0; i < sizeof(DataPacket); ++i) {
-        checksum += bytes[i];
-    }
-    return checksum;
-}
-
-uint64_t generateUniqueID() {
-    srand(static_cast<unsigned>(time(nullptr)));
-    uint64_t id = rand() % 1000 * 1000 + rand() % 1000;
-    return id;
-}
-
-void sendPacket(SOCKET sock, const DataPacket& packet) {
-
-    uint32_t checksum = calculateChecksum(packet);
-
-    DataPacket packetToSend = packet;
-
-    packetToSend.CRCchecksum = checksum;
-
-    send(sock, reinterpret_cast<const char*>(&packetToSend), sizeof(DataPacket), 0);
-}
-
-void receivePacket(SOCKET sock, DataPacket& packet) {
-
-    recv(sock, reinterpret_cast<char*>(&packet), sizeof(DataPacket), 0);
-    uint32_t calculatedChecksum = calculateChecksum(packet);
-    //if (packet.CRCchecksum != calculatedChecksum) {
-    //    std::cerr << "Checksum validation failed" << std::endl;
-    //}
-}
-
-std::string serializePacket(const DataPacket& packet) {
-    std::ostringstream oss;
-    oss.write(reinterpret_cast<const char*>(&packet), sizeof(DataPacket));
-    return oss.str();
-}
-
-DataPacket deserializePacket(const char* buffer) {
-    DataPacket packet;
-    memcpy(&packet, buffer, sizeof(DataPacket));
-    return packet;
-}
 
 void DisplayData(const DataPacket& packet) {
 
     std::cout << std::endl;
-    std::chrono::system_clock::time_point time_point = std::chrono::system_clock::time_point(std::chrono::milliseconds(packet.header.timestamp));
-    std::time_t time = std::chrono::system_clock::to_time_t(time_point);
-    std::ostringstream oss;
-    oss << std::put_time(gmtime(&time), "%Y-%m-%d %H:%M:%S UTC");
-    std::cout << "Timestamp: " << oss.str() << std::endl;
+    getTimeStamp(packet);
     std::cout << "Sequence Number: " << packet.header.sequenceNumber << std::endl;
 
     if (packet.header.packetType == 1) {
